feat(text): change logic to properly use fontsize and update minimap
previously the minimap was hardcoded and thus wouldnt update the size of the text relative to its
real size. Im using a useEffect, which im not happy with. IMO you should not use state to set state,
but at the moment i dont really know what else to do, thus ithis is the solution.